feat(services): implementation of "services" part of the TourOfHeroes
I have generated a service with the CLI and the command `ng g s shared/Hero --flat false`. All the modification where made to use the service from the `toh.module.ts`. The service spec use the fakeAsync logic to test a task which take a certain amount of time.